fxclusterf

+++ src/install.ts
@@ -38,6 +38,7 @@ export const installNotesFromCsv = (csvfile, setname = "FatBoy") => {
     if (name === "") continue;
     const fontname = format(name).replace("\t", "");
     if (!fontname) continue;
+    if (fontname === "Binary_file_standard_input_matches") continue;
     const localname = "mp3/" + setname + "_" + fontname + ".js";
     if (!existsSync(localname)) {
       execSync(
@@ -68,14 +69,14 @@ export const installNotesFromCsv = (csvfile, setname = "FatBoy") => {
       mkfolder(`midisf/${fontname}`);
       const index = midi - 21;
       const pcmname = `midisf/${fontname}/${index}.pcm`;
-      console.log(pcmname);
       try {
-        console.log(
-          `dd if=${localname} bs=${bytesPerNote} skip=${index} count=1 of=pipe:1|ffmpeg -y -hide_banner -loglevel panic -f mp3 -i pipe:0 -f f32le -ac 1 -ar 48000 ${pcmname}`
-        );
-        execSync(
-          `dd if=${localname} bs=${bytesPerNote} skip=${index} count=1 |ffmpeg -y -hide_banner -loglevel panic -f mp3 -i pipe:0 -f f32le -ac 1 -ar 48000 ${pcmname}`
-        );
+        if (!existsSync(pcmname)) {
+          console.log(pcmname);
+
+          execSync(
+            `dd if=${localname} bs=${bytesPerNote} skip=${index} count=1 |ffmpeg -y -hide_banner -loglevel panic -f mp3 -i pipe:0 -f f32le -ac 1 -ar 48000 ${pcmname}`
+          );
+        }
       } catch (e) {
         console.log(e);
       }
@@ -84,36 +85,36 @@ export const installNotesFromCsv = (csvfile, setname = "FatBoy") => {
 };
 // installNotesFromCsv("/home/AzureUser/ssr-bach/csv/serenade_k361_3rd-mid.csv");
 
-if (process.argv[2]) {
-  const csvfile = process.argv[2] || "";
+// if (process.argv[2]) {
+//   const csvfile = process.argv[2] || "";
 
-  const setname = process.argv[3] || "FatBoy";
-  installNotesFromCsv(csvfile, setname);
-}
-for (const localname of readdirSync("./mp3")) {
-  const byteswrote = parseInt(
-    execSync("wc -c mp3/" + localname)
-      .toString()
-      .trim()
-      .split(/\s+/)[0]
-  );
-  const bytesPerNote = ~~(byteswrote / 88 / 4) * 4;
-  const fontname = localname.replace("FatBoy_", "").replace(".js", "");
-  for (let index = 0; index < 88; index++) {
-    //  mkfolder(`midisf/${fontname}`);
+//   const setname = process.argv[3] || "FatBoy";
+//   installNotesFromCsv(csvfile, setname);
+// }
+// for (const localname of readdirSync("./mp3")) {
+//   const byteswrote = parseInt(
+//     execSync("wc -c mp3/" + localname)
+//       .toString()
+//       .trim()
+//       .split(/\s+/)[0]
+//   );
+//   const bytesPerNote = ~~(byteswrote / 88 / 4) * 4;
+//   const fontname = localname.replace("FatBoy_", "").replace(".js", "");
+//   for (let index = 0; index < 88; index++) {
+//     //  mkfolder(`midisf/${fontname}`);
 
-    const pcmname = `midisf/${fontname}/${index}.pcm`;
-    console.log(pcmname);
-    try {
-      console.log(
-        `dd if=mp3/${localname} bs=${bytesPerNote} skip=${index} count=1 of=pipe:1|ffmpeg -y -hide_banner -loglevel panic -f mp3 -i pipe:0 -f f32le -ac 1 -ar 48000 ${pcmname}`
-      );
-      execSync(
-        `dd if=mp3/${localname} bs=${bytesPerNote} skip=${index} count=1 |ffmpeg -y -hide_banner -loglevel panic -f mp3 -i pipe:0 -f f32le -ac 1 -ar 48000 ${pcmname}`
-      );
-    } catch (e) {
-      console.log(e);
-    }
-  }
-  execSync(`mv mp3/${localname} done`);
-}
+//     const pcmname = `midisf/${fontname}/${index}.pcm`;
+//     console.log(pcmname);
+//     try {
+//       console.log(
+//         `dd if=mp3/${localname} bs=${bytesPerNote} skip=${index} count=1 of=pipe:1|ffmpeg -y -hide_banner -loglevel panic -f mp3 -i pipe:0 -f f32le -ac 1 -ar 48000 ${pcmname}`
+//       );
+//       execSync(
+//         `dd if=mp3/${localname} bs=${bytesPerNote} skip=${index} count=1 |ffmpeg -y -hide_banner -loglevel panic -f mp3 -i pipe:0 -f f32le -ac 1 -ar 48000 ${pcmname}`
+//       );
+//     } catch (e) {
+//       console.log(e);
+//     }
+//   }
+//   execSync(`mv mp3/${localname} done`);
+// }
